fix: address review comments for Hytte-z225
Move SpeedPlanSegment and SpeedPlanSegmentKind to the shared
web/src/types/training.ts module so neither component depends on the
other for the domain type. Both SpeedPlanEditor and WorkoutContextModal
now import directly from there. Remove the re-export from
WorkoutContextModal (inverted dependency) and tighten the kind field
from `string` to the `SpeedPlanSegmentKind` union type.
